// src/store/wishlist/Wishlist.types.ts
export enum WISHLIST_ACTION_TYPES {
    CREATE_WISHLIST_ITEM_START = 'wishlist/CREATE_WISHLIST_ITEM_START',
    CREATE_WISHLIST_ITEM_SUCCESS = 'wishlist/CREATE_WISHLIST_ITEM_SUCCESS',
    CREATE_WISHLIST_ITEM_FAILED = 'wishlist/CREATE_WISHLIST_ITEM_FAILED',
    FETCH_WISHLIST_START = 'wishlist/FETCH_WISHLIST_START',
    FETCH_WISHLIST_SUCCESS = 'wishlist/FETCH_WISHLIST_SUCCESS',
    FETCH_WISHLIST_FAILED = 'wishlist/FETCH_WISHLIST_FAILED',
    REMOVE_WISHLIST_ITEM_START = 'wishlist/REMOVE_WISHLIST_ITEM_START',
    REMOVE_WISHLIST_ITEM_SUCCESS = 'wishlist/REMOVE_WISHLIST_ITEM_SUCCESS',
    REMOVE_WISHLIST_ITEM_FAILED = 'wishlist/REMOVE_WISHLIST_ITEM_FAILED',
}